Fix/feat: solved login and registration redirect problems, added logout functionality
